pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voucherify-core-mcp"
dynamic = ["version"]
description = "A Model Context Protocol (MCP) server for integrating with Voucherify services"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Voucherify Dev Team", email = "support@voucherify.io"},
]
maintainers = [
    {name = "Voucherify Dev Team", email = "support@voucherify.io"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Office/Business :: Financial :: Point-Of-Sale",
]
keywords = ["voucherify", "mcp", "model-context-protocol", "coupons", "vouchers", "promotions"]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.8",
    "litellm>=1.74.3",
    "mcp[cli]>=1.11.0",
    "voucherify>=5.0.3",
]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.8",
]

[project.urls]
Homepage = "https://github.com/voucherifyio/core-mcp"
Documentation = "https://github.com/voucherifyio/core-mcp#readme"
Repository = "https://github.com/voucherifyio/core-mcp.git"
Issues = "https://github.com/voucherifyio/core-mcp/issues"
Changelog = "https://github.com/voucherifyio/core-mcp/blob/main/CHANGELOG.md"

[project.scripts]
voucherify-core-mcp = "voucherify_core_mcp.server:main"

[tool.hatch.version]
path = "src/voucherify_core_mcp/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/voucherify_core_mcp"]

[tool.hatch.build.targets.sdist]
exclude = [
    "tests/",
    "playground/",
    "debug.py",
    "vou_*.py",
    "test-*.py",
    "*.pyc",
    "__pycache__/",
    ".pytest_cache/",
    ".git/",
    ".claude/",
    ".cursor/",
    "claude/",
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
